Updated linux + osx
git-svn-id: https://groke.otago.ac.nz/ant/ant@169 7489a7b4-4588-4892-8acd-774bacfab51c

+++ stemmer/ga_stemmer.c
@@ -3,7 +3,7 @@
 #include "ga_stemmer.h"
 #include "counter.h"
 
-long GA_stemmer::stem(char *term, char *destination) {
+size_t GA_stemmer::stem(char *term, char *destination) {
     return strlen(strcpy(destination, individual->apply(term)));
 }
